Fix issues
