+++ package.json
@@ -11,7 +11,7 @@
     "@radix-ui/react-avatar": "^1.1.1",
     "@radix-ui/react-dialog": "^1.1.2",
     "@radix-ui/react-dropdown-menu": "^2.1.2",
-    "@radix-ui/react-icons": "^1.3.1",
+    "@radix-ui/react-icons": "^1.3.2",
     "@radix-ui/react-label": "^2.1.0",
     "@radix-ui/react-popover": "^1.1.2",
     "@radix-ui/react-progress": "^1.1.0",
@@ -24,44 +24,44 @@
     "@radix-ui/react-toast": "^1.2.2",
     "@radix-ui/react-tooltip": "^1.1.4",
     "@supabase/ssr": "^0.5.2",
-    "@supabase/supabase-js": "^2.46.1",
-    "@vercel/analytics": "^1.4.0",
+    "@supabase/supabase-js": "^2.46.2",
+    "@vercel/analytics": "^1.4.1",
     "@vercel/postgres": "^0.10.0",
     "@vercel/speed-insights": "^1.1.0",
-    "class-variance-authority": "^0.7.0",
+    "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
-    "embla-carousel-react": "^8.3.1",
-    "framer-motion": "^11.11.17",
+    "embla-carousel-react": "^8.5.1",
+    "framer-motion": "^11.12.0",
     "geist": "^1.3.1",
-    "lucide-react": "^0.456.0",
+    "lucide-react": "^0.462.0",
     "next": "15.0.3",
     "next-themes": "^0.4.3",
     "react": "19.0.0-rc-65a56d0e-20241020",
-    "react-day-picker": "9.3.0",
+    "react-day-picker": "9.4.1",
     "react-dom": "19.0.0-rc-65a56d0e-20241020",
     "react-feather": "^2.0.10",
     "sharp": "^0.33.5",
     "sonner": "^1.7.0",
-    "tailwind-merge": "^2.5.4",
-    "tailwindcss": "3.4.14",
+    "tailwind-merge": "^2.5.5",
+    "tailwindcss": "3.4.15",
     "tailwindcss-animate": "^1.0.7",
-    "typescript": "5.6.3",
+    "typescript": "5.7.2",
     "use-debounce": "^10.0.4",
     "zod": "^3.23.8"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.2.0",
-    "@eslint/js": "^9.14.0",
-    "@types/node": "22.9.0",
+    "@eslint/js": "^9.16.0",
+    "@types/node": "22.10.1",
     "@types/react": "npm:types-react@19.0.0-rc.1",
     "@types/react-dom": "npm:types-react-dom@19.0.0-rc.1",
     "autoprefixer": "^10.4.20",
     "encoding": "^0.1.13",
-    "eslint": "^9.14.0",
+    "eslint": "^9.16.0",
     "eslint-config-next": "15.0.3",
     "postcss": "^8.4.49",
-    "prettier": "^3.3.3"
+    "prettier": "^3.4.1"
   },
   "pnpm": {
     "overrides": {
